back-end/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
python-dotenv==1.0.0
Flask-JWT-Extended==4.5.2
werkzeug==2.3.7